test(unit): converted the project-type prompt tests to vitest
for #452
