+++ src/project-type-plugin/prompt.js
@@ -1,4 +1,4 @@
-import {prompt, Separator} from '@form8ion/overridable-prompts';
+import {prompt} from '@form8ion/overridable-prompts';
 
 import {questionNames} from '../prompts/question-names.js';
 
@@ -9,7 +9,7 @@ export default async function ({types, projectType, decisions}) {
     name: questionNames.PROJECT_TYPE_CHOICE,
     type: 'list',
     message: `What type of ${projectType} is this?`,
-    choices: [...Object.keys(types), new Separator(), 'Other']
+    choices: [...Object.keys(types), 'Other']
   }], decisions);
 
   return answers[questionNames.PROJECT_TYPE_CHOICE];
